requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_tflow_select=2.3.0=mkl
absl-py=0.13.0=py38h06a4308_0
aiohttp=3.7.4=py38h27cfd23_1
astor=0.8.1=py38h06a4308_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chardet=3.0.4=py38h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
configparser=5.0.2=pypi_0
coverage=5.5=py38h27cfd23_2
cryptography=3.4.7=py38hd23ed53_0
cudatoolkit=11.1.1=h6406543_8
cython=0.29.24=py38h295c915_0
decorator=5.0.9=pyhd3eb1b0_0
docker-pycreds=0.4.0=pypi_0
ffmpeg=4.2.2=h20bf706_0
freetype=2.10.4=h5ab3b9f_0
gast=0.4.0=pyh9f0ad1d_0
gitdb=4.0.7=pypi_0
gitpython=3.1.18=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.36.1=py38h2157cd5_1
h5py=2.10.0=py38hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py38h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.26.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py38h06a4308_1
jpeg=9b=h024ee3a_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py38h06a4308_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
multidict=5.1.0=py38h27cfd23_2
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
opt_einsum=3.3.0=pyhd3eb1b0_1
parso=0.8.2=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py38h2c7a002_0
pip=21.2.2=py38h06a4308_0
promise=2.3=pypi_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.17.2=py38h295c915_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pysocks=1.7.1=py38h06a4308_0
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd3eb1b0_0
pytorch=1.9.0=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.2.1=py38h295c915_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.6.2=py38had2a1c9_1
sentry-sdk=1.3.1=pypi_0
setuptools=52.0.0=py38h06a4308_0
shortuuid=1.0.1=pypi_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
subprocess32=3.5.4=pypi_0
tensorboard=2.5.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.4.1=mkl_py38hb2083e0_0
tensorflow-base=2.4.1=mkl_py38h43e0292_0
tensorflow-estimator=2.5.0=pyh7b7c402_0
termcolor=1.1.0=py38h06a4308_1
tk=8.6.10=hbc83047_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu111
tornado=6.1=py38h27cfd23_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
wandb=0.12.0=pypi_0
wcwidth=0.2.5=py_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_0
wrapt=1.12.1=py38h7b6447c_1
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
